enterprise_ai/requirements.txt
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.13
langchain-core==0.1.10
langchain-text-splitters==0.0.1
langchain-chroma==0.0.1
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
chromadb==0.4.18
faiss-cpu==1.7.4
redis==5.0.1
tiktoken==0.5.1
pillow==10.1.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
milvus-client==2.3.1
opencv-python==4.8.1.78 